updated nuFac stuff

+++ nuF.h
@@ -3,6 +3,6 @@
 #define NUF_H
 
 void nuF(double * nuFac, double nu, int n, double vthe, 
-	 double delV, double Nvsteps, char * scattType);
+	 double delV, int Nvsteps, char * scattType);
 
 #endif
